requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
python-dotenv==1.0.0
werkzeug==3.0.1
nfl-data-py==0.3.2
bcrypt==4.0.1
psycopg2-binary==2.9.9
redis==5.0.1